stats and split files
