+++ analysis/util.py
@@ -1,0 +1,43 @@
+import os
+
+def read_output_files_and_perform(func, *args):
+	os.chdir('../outputs/')
+	languages, options = parse_output_filenames()
+	
+	for option in options:
+		numbers = option[0]
+		trials = option[1]
+		func(languages, numbers, trials, *args)
+
+def parse(line):
+	s = line.split(":")
+	
+	n = float(s[0].strip())
+	probability = float(s[1].strip())
+
+	return [n + 1, probability]
+
+def parse_output_filenames():
+	'''
+	 Expecting to be within /outputs/
+	'''
+	ls = os.listdir()
+
+	languages = set()
+	options = set()
+	for file in ls:
+		if '.' in file: continue
+
+		if "_" in file:
+			fileparts = file.split("_")
+			
+			language = fileparts[0]
+			numbers = int(fileparts[1])
+			trials = int(fileparts[2])
+
+			option_tuple = (numbers, trials)
+
+			languages.add(language)
+			options.add(option_tuple)
+
+	return (languages, options)
